Refactor: Add new functions for document and linked model retrieval in query module Renamed function in query preflight module and made a count module under pyrevit.revit.db Refactor audit_all accordingly
